build: the last sign-conversion, which missed its commit
PARAM_NAMES is indexed by an int32_t counter. It belongs with 7a9a9f3, whose
`git add` covered src/ and include/ but not ml/cpp -- so the zero-warning
census that commit reports was measured with this fix present in the working
tree and absent from the commit. The census is honest now.

+++ ml/cpp/GeneticAlgorithm.cpp
@@ -238,7 +238,7 @@ void printAsCppInitializer(const Individual& individual, const char* label) {
 
     // Also print named parameters
     for (int32_t i = 0; i < NUM_PARAMS; ++i) {
-        std::fprintf(stderr, "  %-30s = %.3f\n", PARAM_NAMES[i],
+        std::fprintf(stderr, "  %-30s = %.3f\n", PARAM_NAMES[static_cast<std::size_t>(i)],
                      static_cast<double>(individual.genes[static_cast<std::size_t>(i)]));
     }
 }
